feat: refactor by pr comments
