+++ airways/src/environments/environment.ts
@@ -4,4 +4,5 @@ export const environment = {
     'https://autocomplete.travelpayouts.com/places2?locale=en&types=airport&term=',
   autocompleteApi:
     'https://autocomplete.travelpayouts.com/places2?locale=en&types[]=airport&types[]=city&term=',
+  base_url: 'http://localhost:3000/',
 };
